refactor: persistence 패키지를 도메인별로 재구조화
- UserBalance 관련 파일들을 user 패키지로 이동
- 일관된 패키지 구조 적용 (도메인별 분리)
